refactor: move the round's additions out of two capped files
`check_source_size_budget` charges every line above 1500 per file, and
this round's additions put git_authorization_hook.ts at 1601 and
dispatch_hook.ts at 1589 — 195 new violations against an 18068 baseline
that only turns one way.

Two extractions along real seams, not line-shuffling. The hook owns the
machinery — sentence bounds, negation, fenced-paste handling — while the
consequence vocabulary and its object extraction move to
_lib/consequence_objects.ts. The dispatcher runs concerns and reduces
exit codes, while the four-condition permission decision moves to
hooks/permission_gate.ts.

The extraction also removed a duplicate: `_payload_tool_name` briefly
existed in both modules, which is the drift host_semantics warns about
for its own predicate. One copy now, re-exported so the public surface
is unchanged.

The doctor check is re-homed with the runtime-wiring family rather than
carrying its own registration. That is the better placement on the
merits — `checkHookWiring` already reads the same ~/.claude/settings.json
for the same host — and it costs cmd_doctor.ts, which is 2100 lines past
the cap, zero new lines.

One pre-existing markdown heading in a dispatch_hook doc comment came
into the comment gate's window with the extraction and is flattened.
